4094: add note about chop
While checking this I noticed that chopping occurs for
any number in an expression, not only the final result.
This is sub-optimal. e.g. (pi/65).n(chop=.01) rounds to
zero because the 1/65 rounds to less than 0.02 and is
replaced with zero even though if the rounding was delayed
until the multiplication by pi occured, it wouldn't have
been rounded. I did not clutter the docs with this detail, however.