4107: Add _eval_Eq to MatrixExpr.
Replaces the Equality with an And containing one Equality for each element of
the matrix.  Anything with a different shape, or no shape attribute, will
evaluate as False.  This only applies to immutable matrices, but Equality
automatically converts any input matrices to immutable ones anyway, so this
will not present an issue.