make nargs of functions with args an int
nargs, defined in classes like sin and cos are integers, reflecting
the fact that they take only one argument; in the case of log, the
class defines nargs as a tuple since log can take 1 or 2 arguments.
For consistency, at instanciation all nargs are converted to a tuple
of acceptable values. If any number of arguments are acceptable, then
the empty tuple is used.

    f(x).nargs == sin(x).nargs == log(x).nargs == 1

To set the number of args for user-defined function or WildFunction,
use keyword `nargs`, setting it to the desired value (or list of
acceptable values).