Improvements to IPython printing: strings, floats, and ints.
Removed strings from LaTeX printing in IPython as this was causing dictionaries
that contained strings and other valid LaTeX printable types to render. This is
typically not desirable especially if you have a dictionary with only strings,
floats, and integers.

I also added an option `print_builtin` to `init_printing` that allows you to
disable the float and integer LaTeX printing. For example, if you don't want a
dictionary of integer keys and float values to render in LaTeX.

Fixes issue #2672.