Add a fact about primes
There's not much we can do here without inequality assumptions, because we
need greater than 1, not greater than 0, for most assumptions.

Also found an oddity: 1 is considered to be composite, even though it usually
isn't. This is no doubt due to the lack of inequality assumptions, but it
makes certain facts impossible, such as composite*prime => !prime, (e.g., 1*2
= 2 is prime).

Also, this is wrong in the current new assumptions (they assume that the
product of *any* two integers is not prime).