Set RR.get_ring() to RR
Mathematically, it doesn't make sense to ask what "the ring" associated with a
field is, since every field is itself a ring, and most have several
non-trivial subrings, so I'm not sure what this function is trying to
achieve. This seems to work. This fixes unification of things like RR[x] and
ZZ(y), which would result in DomainError: there is no ring associated with
RR.