Fixes issue #2769
The translate function in the latex printing module was not being called on
single variables in vector expressions. This is now implemented in the
mechanics latex printer so that variables show up correctly in Latex printing.
Note that gamma is reserved for gamma functions in SymPy and is not considered
a valid variable name for simple translation to its latex form.