change elif to if and other edits
The two sequential ifs will handle the case when
p = self.center of the ellipse.

doctests were edited to reflect current output

more transparent expressions were used to calculate
the intersections (slope=oo is more readily seen to
be a vertical line)

The code is more ready for the arbitrary ellipse to
be introduced.