give normal_lines a numerical option
Also, all solutions have to be checked since if None is
returned in checksol for a test of assumptions, e.g. that
a solution is real, then that solution will be returned.
We don't want that here so we check the numerical value
but keep the symbolic one (unless prec is given).