block imaginary coordinates for Point
The computations that are done in geometry assume that
the points are real. Without placing undue stress on the
computations by checking this every time, a check is done
only when the coordinates are numbers. Of course, there
is nothing to stop one from using imaginary points in
an expression obtained from geometry, but one should not
be doing that.