fix AskPositiveHandler.Pow
There was a duplicate .even test; the second one should have
been odd. It was not caught the last time changes were made
to these lines, probably because the change was automated. I
found the need for a change because this was a line that was
uncovered with tests.

Other tests were also added to bring the coverage in `sets` almost
to 100% (the handling in _number is not complete but may represent
lines that will be unreachable/unlikely). In the `order` file there
is a significant amount of code for the hermitian property that
is not covered.