Changed lambdastr to dummy symbols if dummify=True
Previously, if the user declared alpha = symbols('\alpha'),
lambdastr would try to use \alpha as a variable. Now all
all symbols are replaced with dummies if dummify is true,
so any string is a valid variable name.

If the user wants to not use Dummy symbols, they can still
set dummify=False, and get direct substitution. This allows
the output to be examined, but will result in a syntax error
in lambdify if a non-valid symbol name is used (such as \alpha)