Rename subset() to is_subset() and add is_superset() #7460
The subset() method is now marked as deprecated. The is_subset() method
also has a different semantic meaning now.

Earlier, self.subset(other) checked if 'other' is a subset of
'self'. However, now with the renaming it makes more sense that the
method call self.is_subset(other) is interpreted as check if 'self' is a
subset of 'other'.

A corresponding is_superset() method has been added to allow checking if
one set is a superset of another.