ODE:Corrected classify_ode of separable_reduced Issue #6950
There were few equations like sin(x)*cos(f(x)) + cos(x)*sin(f(x))*f(x).diff(x) which separable_reduced classified but was not able to solve because it was not fitting in its general form. So, this patch modifies classify_ode of separable_reduced so as not to classify those equations which are out of its scope.