Latex format
Ok no problem it's back on Latex.  As for the accent, it seems like the entire doc should either be with it or without it, not both.  As it is, the accent was only used in about 1 or 2 places and not used almost everywhere else.  Some sites use it, some don't.