Fix unit system example with solve
The documentation has an example on how to use solve with the unitsystem
module, but there was difficulties due to Dummy argument. As a (temporary)
solution the system is solved before replacing with units.
As a side effect I had to allow non-integer powers for dimensions, since they
may appear in intermediate steps.