Check for system of eq being solvable under classify_sysode
This PR contain methods to classify whether a given system of equation
is solvable by current solver. It is broadly divided on the basis of
parameters returned by classify_sysode which are linearity, no of eq
and order. If the return parameter by these methods is any "type" then
the given system is solvable otherwise None is returned which means
it is not solvable by current solving engine.