Started work on Linearizer Class
Implemented routine for constraints on position and velocity.
Unsure if it will work if either are missing, most likely will need
to add control flow for this. The API is also complicated, and will
be simplified as more code is ported over.

Also added one test for the new linearize routine. Based off of
the example in Luke and Gilbert's paper. As more functionality is
added to the linearization routine, the work done in this test
will get smaller. Still need to add more tests from the examples
I worked out last week.