Use sorted Omega list, ordered by default_sort_key
Before last element of Omega was chosen as "w" after sorting it
based on node.ht().  However in some cases, all the elements of
Omega had the same value of node.ht().

So Omega is sorted based on the default_sort_key before so that its
ordering is not random.