distribute dummy on Add in _osborne
Rather than make the osborne and the osbornei roundtrip by dividing
by d in osbornei to undo the multiplying in osborne, the simple
replacement of d with 1 is used instead.

If the dummy is not distributed, arguments that looked like the addition
of two terms no longer appear so and some transformations (like TR12)
then fail.