Changed lambdify to use numpy.array for Matrix
This is the first step on moving towards using numpy.array to represent
sympy.Matrix by default. This change adds a kwarg `new_defaults` that
defaults to False. If it's False, the previous behavior (numpy.matrix)
is used, and a SymPyDeprecationWarning is issued. If it's True, the new
default of numpy.array is used. In the next release this kwarg should be
removed.