Relational: Unify `Lt` and `<`, a fix for pr #7783
Make long form and symbolic operators take the same path through the code.
This effects `Lt`, `Gt`, `Ge`, `Le`, their longer forms (`StrictLessThan`,
etc) and the symbolic versions (`<`, etc) which are usually accessed as
`__lt__` etc.

Here is an outline of the approach:

1. If we come via `__lt__` than we do all we can to reduce to bool or
raise exception.  This can go through several classes' `__lt__` methods,
but in the end (currently `Expr.__lt__`) we ask `Lt` to make a
non-evaluated inequality with `evaluated=False`.

2. On the other hand, if we come from `Lt(lhs, rhs)`, we first ask
`lhs.__lt__(rhs)`.  Which will then either return a boolean, raise an
exception or eventually call `Lt` again but with `evaluated=False` as
above.

In this fashion, we always have the same result whichever form we called.