modify test which takes a long time
This only takes a long time in this branch because an attempt
at evaluating the integral is attempted now...and that integral
is zero and this takes a long time in the evalf routine. Since
we are only testing the type of arguments in this test (not
the evaluation of the integral) it seems reasonable to change the test.