7895: watch out for new atan2 in solve
When accepting real and imaginary parts of an equation, discriminate
against additional atan2 terms, too. I'm not sure that recording what
was already present is necessary, but it seems like a good idea to
not reject a splitting into real and imag as long as no *new* re, im,
arg or atan2 are introduced.

The test for issue 6121 was changed to a simple assertion since the
original issue involved an infinite recursion. The new result is
a little different than the old result but there is no error.