Proof of concept: new assumptions check old assumptions for Symbol
That is, ask(Q.real(Symbol('x', real=True))) now gives True (but it only
checks old assumptions Symbols, not all expressions).

They only check for Symbol to reduce the number of test failures and things
that need to be fixed. So far, only one thing fails, which is
ask(Q.commutative(x), ~Q.commutative(x)), because x is commutative by default
in the old assumptions. If we go forward with this, we should probably either
have the new assumptions override the old ones, or else disallow them to
conflict.