Relational: don't use symbol ops `<` in `Lt` lest it reorder arguments
I encountered and filed issue #7951, where `x < p` can become `p > x`
even when both are sympified.  This adds tests.  Issue causes failures
in `meijerint._condsimp` and more specifically `subs` where an `cls=Wild`
variable was substituted for a regular symbol in an Inequality.

As a workaround (which is probably safer against this sort of bug), we
call `__lt__` and friends instead of the symbols.