Relational: if simplify lhs/rhs not enough, try harder with equals(0)
Call `_eval_sides` when simplifying inequalities (this can be expensive
so we first try the existing simplify lhs/rhs approach).

A consequence of unifying `Lt` and `<` (the fix for pr #7783) is that
`_eval_sides` is no longer called in `Relational.__new__`.  Sometimes
this meant inequalities did not reduce to boolean when they did before.
Now we have the essentially the same code: if an expression previously
reduced to bool upon creation, it should now do so after simplify().