do not call self.is_* from _eval_is_* routines
I am not sure about the ramifications but my hunch is that
if, when trying to resolve fact foo, an _eval_is routines
calls self.is_bah -- and bah is not actually a prerequisite
of foo -- that this might interrupt the resolution of fact
foo. At least, using the _eval method will result in a faster
resolution of the fact without running through prerequisites
of the fact. But again...I'm not even sure that's a good idea.
I'm not sure how much of the burden of knowing what to call
should be the responsibility of the _eval_is routines and how
much should be handled by the assumption machinery. (I know the
machinery will handle prerequisites, but it is not clear how
to effectively write the _eval_is routines so they are not
working too hard.