Old assumptions: replace !finite == infinite with infinite -> !finite
The motivation here is that there might be things that are neither finite nor
infinite, i.e., non-numbers. For example, I would expect anything with
commutative=False to be neither finite nor infinite.

This only changes the fact in assumptions.py, and does not fix any of the
remaining code in SymPy to match it.