Old assumptions: change nonzero == !zero to nonzero == real & !zero
This matches the definition in the new assumptions. This definition is more
useful, as !zero can just as easily be accessed by not expr.is_zero. This also
matches the other non- assumptions, namely nonnegative and nonpositive, which
mean real & !negative and real & !positive, respectively. This definition may
end up being a "gotcha", but hopefully its usefulness will outweigh this.