Old assumptions: Add the fact complex -> finite
The only reason I didn't add the fact complex == finite is that it seems to
make finite pointless, as it would become an exact alias for complex. However,
I am not quite sure what would be finite and not complex. Perhaps we should
make it this, or even delete finite and just use complex.

Note that this will require changing things like oo.is_real == True, as this
makes all complex numbers finite. I have not made any changes outside of
assumptions.py yet.  It will likely be necessary to add extended_real and/or
extended_complex assumptions as aliases for real | infinite and complex |
infinite.