Add extended_real to the old assumptions
This only adds the fact to assumptions.py. It does not change any of the
corresponding code elsewhere.

I have made the change that positive and negative are the extended versions
(i.e., oo and -oo are positive and negative).  This means that real is finite,
and positive does not imply real.

I did not make nonzero or noninteger imply extended_real, but perhaps I
should, at least for nonzero.