do not use self._eval_is from within _eval_is routines
When _eval calls are made they short circuit the checks
and inferences made by the assumptions engine. If, instead
of this change, you changed the few instances in _eval_is_* routines
of self.is_* to self._eval_is_* there will be a failure in the
meijerint.py file in test_issue_6348.