modify roots_binomial and roots_quadratic
These two routines supply roots to _roots_trivial which expects
the results to be in the same order as the RootOf instances. Although
the sorting could take place in _roots_trivial, all the information
about the roots is in these routines so it makes sense to sort them
in that information-rich context. In order to keep this fast, no numerical
calculations are made.

Sorting is a slow operation and was added to get consistent results
from these routines (all roots_foo in polyroots). Now that Add's args
are not sorted by hash this is no longer an issue, so sorting is
now a private option to these private functions but is left True.

TODO remove sorting from all routines (issue 8294).

An advantage to not sorting is the freedom to fail to return items
in any sorted order. Nonetheless, when the domain is ZZ the roots
will be sorted (reals first, then non-reals according to real then
imaginary parts) and that is what is important to _roots_trivial.